+++ apps/api/src/modules/finance/finance.module.ts
@@ -1,7 +1,8 @@
 import { Module } from '@nestjs/common';
 import { FinanceController } from './finance.controller';
 import { FinanceService } from './finance.service';
+import { KpisModule } from '../kpis/kpis.module';
 
-@Module({ controllers: [FinanceController], providers: [FinanceService], exports: [FinanceService] })
+@Module({ imports: [KpisModule], controllers: [FinanceController], providers: [FinanceService], exports: [FinanceService] })
 export class FinanceModule {}